requirements/base.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
